+++ src/lib.rs
@@ -87,13 +87,14 @@ pub use glam::{DMat3, DMat4, DQuat, DVec2, DVec3, DVec4};
 
 #[cfg(feature = "geo")]
 pub use geo::{
-    Aabb, Bvh, Capsule, ContactConstraint, ConvexDecomposition, ConvexHull3D, ConvexPolygon,
-    ConvexSupport, ConvexSupport3D, Frustum, KdTree, Line, Obb, Octree, Penetration3D, Plane,
-    Quadtree, Ray, Rect, Segment, SpatialHash, Sphere, TriMesh, Triangle,
+    Aabb, Bvh, Capsule, ContactConstraint, ContactEdge, ConvexDecomposition, ConvexHull3D,
+    ConvexPolygon, ConvexSupport, ConvexSupport3D, Frustum, HalfEdge, HalfEdgeMesh, Island, KdTree,
+    Line, Obb, Octree, Penetration3D, Plane, Quadtree, Ray, Rect, Segment, SpatialHash, Sphere,
+    TriMesh, Triangle,
 };
 
 #[cfg(feature = "num")]
-pub use num::{Complex, CsrMatrix, EigenDecomposition, OptResult, Pcg32, Svd};
+pub use num::{Complex, CsrMatrix, DenseMatrix, EigenDecomposition, OptResult, Pcg32, Svd};
 
 #[cfg(feature = "autodiff")]
 pub use autodiff::{Dual, Tape, Var};